backend/requirements.txt
# Core Django & DRF
Django
djangorestframework
django-cors-headers

# Authentication
djangorestframework-simplejwt

# Database
psycopg2-binary

# Environment
python-decouple

# API Documentation
drf-spectacular

# Background Jobs
celery
redis
django-celery-beat
django-celery-results

# Development
django-debug-toolbar

# Utilities
Pillow
django-filter
django-extensions

# Security
django-ratelimit

# Testing
factory-boy
pytest
pytest-django